Represent PDUs as numba jitclasses
This allows fast and convenient access to PDU contents
from njit compiled code.

Accessing a jitclass from normal Python code, however, seems a bit
slower than using a normal class or a dict.
Depending on how much of the L2 code is going to be njit compiled,
this may or may not be a useful approach.
If it turns out to be not useful, the jitclasses could be just
changed to normal classes later.